+++ packages/conan/recipes/openexr/conanfile.py
@@ -2,7 +2,7 @@
 # Copyright (c) Contributors to the aswf-docker Project. All rights reserved.
 # SPDX-License-Identifier: MIT
 #
-# From: https://github.com/conan-io/conan-center-index/blob/cceee569179c10fa56d1fd9c3582f3371944ba59/recipes/openexr/3.x/conanfile.py
+# From: https://github.com/conan-io/conan-center-index/blob/fcf375e3136091137a012a00fab949694edc2da6/recipes/openexr/3.x/conanfile.py
 
 from conan import ConanFile
 from conan.errors import ConanInvalidConfiguration
@@ -12,7 +12,7 @@ from conan.tools.files import apply_conandata_patches, export_conandata_patches,
 from conan.tools.scm import Version
 import os
 
-required_conan_version = ">=1.53.0"
+required_conan_version = ">=2.0"
 
 
 class OpenEXRConan(ConanFile):
@@ -33,20 +33,6 @@ class OpenEXRConan(ConanFile):
         "shared": False,
         "fPIC": True,
     }
-
-    @property
-    def _min_cppstd(self):
-        if Version(self.version) >= "3.3":
-            return 17
-        return 11
-
-    @property
-    def _minimum_compiler_version(self):
-        return {
-            "17": {
-                "gcc": "9"
-            }
-        }.get(str(self._min_cppstd), {})
 
     @property
     def _with_libdeflate(self):
@@ -84,12 +70,7 @@ class OpenEXRConan(ConanFile):
             self.requires("openjph/0.24.5")
 
     def validate(self):
-        if self.settings.compiler.get_safe("cppstd"):
-            check_min_cppstd(self, self._min_cppstd)
-
-        minimum_version = self._minimum_compiler_version.get(str(self.settings.compiler))
-        if minimum_version and Version(self.settings.compiler.version) < minimum_version:
-            raise ConanInvalidConfiguration(f"{self.ref} requires {self.settings.compiler} >= {minimum_version}")
+        check_min_cppstd(self, 11)
 
     def source(self):
         get(self, **self.conan_data["sources"][self.version], strip_root=True)
@@ -141,17 +122,11 @@ class OpenEXRConan(ConanFile):
     def _add_component(self, name):
         component = self.cpp_info.components[self._conan_comp(name)]
         component.set_property("cmake_target_name", f"OpenEXR::{name}")
-        component.names["cmake_find_package"] = name
-        component.names["cmake_find_package_multi"] = name
         return component
 
     def package_info(self):
         self.cpp_info.set_property("cmake_file_name", "OpenEXR")
         self.cpp_info.set_property("pkg_config_name", "OpenEXR")
-
-        self.cpp_info.names["cmake_find_package"] = "OpenEXR"
-        self.cpp_info.names["cmake_find_package_multi"] = "OpenEXR"
-        self.cpp_info.names["pkg_config"] = "OpenEXR"
 
         lib_suffix = ""
         if not self.options.shared or self.settings.os == "Windows":